+++ CHANGELOG.md
@@ -11,6 +11,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/) 
 - Add 'hersens' as alternative spelling in Dutch for 'brain'. Fixes [#1169](https://github.com/fniessink/toisto/issues/1169).
 - When giving the meaning of incorrect answers, don't repeat labels that are homographs. Fixes [#1174](https://github.com/fniessink/toisto/issues/1174).
 - Add 'wateren' as alternative spelling in Dutch for 'waters'. Fixes [#1175](https://github.com/fniessink/toisto/issues/1175).
+- Cloze test quizzes would get a useless tip when presented multiple times. Fixes [#1176](https://github.com/fniessink/toisto/issues/1176).
 - Limit the number of examples shown to three per quiz. Fixes [#1180](https://github.com/fniessink/toisto/issues/1180).
 
 ## 0.40.0 - 2025-09-28